Move FantasyStage type to separate types file and export interfaces
